Refactor Year to model package. Add private isValidYear method Edit SessionDate Message_constraints
