+++ src/main/java/seedu/address/logic/commands/GetMonthlyFeeCommand.java
@@ -13,8 +13,8 @@ import seedu.address.commons.util.DateUtil;
 import seedu.address.commons.util.FeeUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.Year;
 import seedu.address.model.fee.Month;
-import seedu.address.model.fee.Year;
 import seedu.address.model.student.Name;
 import seedu.address.model.student.Student;
