+++ src/main/java/seedu/address/model/fee/MonthlyFee.java
@@ -3,6 +3,8 @@ package seedu.address.model.fee;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
+import seedu.address.model.Year;
+
 /**
  * Represents the monthly fee of a particular month and year that is used for the UI.
  */
