+++ src/test/java/seedu/address/model/YearTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.fee;
+package seedu.address.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
