+++ src/test/java/seedu/address/model/fee/MonthlyFeeTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.Year;
+
 public class MonthlyFeeTest {
 
     private double validMonthlyFee = 30.20;
